mapSetup: unsigned shift for getDataFileSize bound check
Cast getDataFileSize result to u32 before >>5 so the bound check emits
srwi (logical) matching target, not srawi (arithmetic). 89.87->90.49.

+++ src/main/shader.c
@@ -845,7 +845,7 @@ void mapSetup(int mapType, s32* outMapId, s32* outEvent, f32 a, f32 b, f32 c)
     curMapLayer = 0;
     mapY = (s32)fastFloorf(c / gMapBlockWorldSize);
     mapId = mapCoordsToId((s32)fastFloorf(a / gMapBlockWorldSize), mapY, layer);
-    if (mapId < 0 || mapId >= (getDataFileSize(0x1f) >> 5))
+    if (mapId < 0 || mapId >= (s32)((u32)getDataFileSize(0x1f) >> 5))
     {
         curMapType = 0;
     }
